安定性に関する複数の問題を修正
* 連続して通信を受け付けるとサーバーが落ちる
* 爆弾が正しくキューに追加されない
なお, 性能検査は必要.
