+++ src/srv/srv.c
@@ -20,27 +20,6 @@ void spawn_slaves();
 int srv_init();
 void usage();
 
-int recv_magic(int fd)
-{
-	int magic;
-	int rlen;
-	rlen = recv(fd, &magic, sizeof(int), 0);
-	switch (magic) {
-		case PLA:
-			/* Fall-through */
-		case BOM:
-			/* Fall-through */
-		case WAL:
-			/* Fall-through */
-		case END:
-			return magic;
-			break;
-		default:
-			fprintf(stderr, "(Master) Not magic: %x\n", magic);
-			return -1;
-	}
-}
-
 int main(int argc, char **argv)
 {
 	int srv_loop_ret;
@@ -79,133 +58,6 @@ int main(int argc, char **argv)
 		printf("Server started as pid: %d\n", pid);
 		exit(0);
 	}
-}
-
-int recv_single_bomb(int fd, struct bomb *buf, int *id)
-{
-	int recv_len;
-	int len = 0;
-	int total_len = 0;
-	char *head;
-	head = (char *)id;
-	while (len < sizeof(int)) {
-		if ((recv_len = recv(fd, head, sizeof(int) - len, 0)) == -1)
-			return -1;
-		len += recv_len;
-		head += recv_len;
-	}
-	total_len += len;
-
-	head = (char *)buf;
-	len = 0;
-	while (len < sizeof(struct bomb)) {
-		if ((recv_len = recv(fd, head, sizeof(struct bomb) - len, 0)) == -1)
-			return -1;
-		len += recv_len;
-		head += recv_len;
-	}
-	total_len += len;
-
-	return total_len;
-}
-
-int recv_single_player(int fd, struct player *buf, int *id)
-{
-	int recv_len;
-	int len = 0;
-	int total_len = 0;
-	char *head;
-	head = (char *)id;
-	while (len < sizeof(int)) {
-		if ((recv_len = recv(fd, head, sizeof(int) - len, 0)) == -1)
-			return -1;
-		len += recv_len;
-		head += recv_len;
-	}
-	total_len += len;
-
-	head = (char *)buf;
-	len = 0;
-	while (len < sizeof(struct player)) {
-		if ((recv_len = recv(fd, head, sizeof(struct player) - len, 0)) == -1)
-			return -1;
-		len += recv_len;
-		head += recv_len;
-	}
-	total_len += len;
-
-	return total_len;
-}
-
-int recv_single_wall(int fd, struct wall *buf, int *id)
-{
-	int recv_len;
-	int len = 0;
-	int total_len = 0;
-	char *head;
-	head = (char *)id;
-	while (len < sizeof(int)) {
-		if ((recv_len = recv(fd, head, sizeof(int) - len, 0)) == -1)
-			return -1;
-		len += recv_len;
-		head += recv_len;
-	}
-	total_len += len;
-
-	head = (char *)buf;
-	len = 0;
-	while (len < sizeof(struct bomb)) {
-		if ((recv_len = recv(fd, head, sizeof(struct bomb) - len, 0)) == -1)
-			return -1;
-		len += recv_len;
-		head += recv_len;
-	}
-	total_len += len;
-
-	return total_len;
-}
-
-int send_end(int fd, int id)
-{
-	int end = END;
-	if (send(fd, &end, sizeof(end), 0) == -1)
-		return -1;
-	if (send(fd, &id, sizeof(int), 0) == -1)
-		return -1;
-}
-
-int send_single_bomb(int fd, struct bomb *b, int id)
-{
-	int  bom = BOM;
-	if (send(fd, &bom, sizeof(int), 0) == -1)
-		return -1;
-	if (send(fd, &id, sizeof(int), 0) == -1)
-		return -1;
-	if (send(fd, b, sizeof(struct bomb), 0) == -1)
-		return -1;
-}
-
-int send_single_player(int fd, struct player *p, int id)
-{
-	int pla = PLA;
-	if (send(fd, &pla, sizeof(int), 0) == -1)
-		return -1;
-	if (send(fd, &id, sizeof(int), 0) == -1)
-		return -1;
-	if (send(fd, p, sizeof(struct player), 0) == -1)
-		return -1;
-	return 0;
-}
-
-int send_single_wall(int fd, struct wall *w, int id)
-{
-	int  bom = BOM;
-	if (send(fd, &bom, sizeof(int), 0) == -1)
-		return -1;
-	if (send(fd, &id, sizeof(int), 0) == -1)
-		return -1;
-	if (send(fd, w, sizeof(struct bomb), 0) == -1)
-		return -1;
 }
 
 void spawn_slaves()
